requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.3
django-cors-headers==2.4.0
-e git+https://github.com/django-import-export/django-import-export.git@596a4f5b54b7d0beadc588b5f59068ea30ed072a#egg=django_import_export
django-mathfilters==0.4.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
fpdf==1.7.2
html5lib==1.0.1
httpie==1.0.2
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.5.12
pdfkit==0.6.1
Pillow==5.3.0
pycparser==2.19
Pygments==2.3.1
Pyphen==0.9.5
pytz==2018.7
PyYAML==3.13
reportlab==3.5.12
requests==2.21.0
six==1.12.0
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
urllib3==1.24.1
WeasyPrint==43
webencodings==0.5.1
wkhtmltopdf==0.2
xlrd==1.2.0
xlwt==1.3.0
